Split test_isbn.py from test_util.py
